Create utils/users module
